4-2_api_basics/requirements.txt
# ChatGPT API
openai==1.3.0

# Google API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gspread==5.12.0

# 環境変数管理
python-dotenv==1.0.0

# HTTP リクエスト
requests==2.31.0

# LINE Messaging API
line-bot-sdk==3.5.0